requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
coverage==6.5.0
Django==3.2.15
djangorestframework==3.14.0
idna==3.4
pkg_resources==0.0.0
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.3
typing_extensions==4.3.0
urllib3==1.26.12
